pyproject.toml
[tool.poetry]
name = "+"
version = "0.1.0"
description = ""
authors = ["Robert Meissner <robert.meissner@codecentric.de>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
datasets = "^2.5.1"
setuptools = "^65.4.1"
openai = "^0.23.1"
python-dotenv = "^0.21.0"
spacy = "^3.4.1"
keybert = "^0.6.0"
nltk = "^3.7"
nlpcloud = "^1.0.32"
scipy = "1.7.2"
pre-commit = "^2.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
